<!-- 🎉🎉🎉 Thank you for the PR!!! 🎉🎉🎉 -->


perf improvement. check if a file is changed in gowalkdir.