@tejal29 2020-06-06
Performance improvements
```

Number of Files | Total Build Time Full Snapshot - current | Total Build Time -  Walk performance | % decrease
-- | -- | -- | --
10000 | 13.834973 | 13.579324 | 1.85%
50000 | 47.926338 | 42.460665 | 11.40%
100000 | 96.440554 | 77.774629 | 19.35%
200000 | 187.346273 | 175.198485 | 6.48%
300000 | 341.233622 | 289.063343 | 15.29%
500000 | 1404.410121 | 1122.947834 | 20.04%
700000 | 1772.379601 | 1350.19682 | 23.82%

```